backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.81.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.129.0
fonttools==4.61.1
greenlet==3.3.1
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
kiwisolver==1.4.9
matplotlib==3.10.8
mysql-connector-python==9.7.0
numpy==2.4.0
openai==2.21.0
packaging==25.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
